requirements.txt
django-bingo
tweepy
